requirements.txt
flask-ngrok
joblib
nltk
pandas
scikit-learn
